fix(webserver): mark PagedResults.total as required
PR #17366 boxed `total` to `Long` without validation when introducing
cursor pagination, so every offset-mode list endpoint (which always
sets it) lost its OpenAPI "required" contract — the generated SDK
types made `total` optional everywhere, breaking downstream TypeScript
consumers that assign it straight to a non-optional field.

Restore the `@NotNull` contract on `PagedResults.total` for the
offset-only endpoints (the vast majority), and split the log-search
cursor path (which can legitimately omit `total`) into its own
`CursorPagedResults` type so that corner case no longer has to weaken
the shared schema.
