+++ webserver/src/test/java/io/kestra/webserver/controllers/api/LogControllerTest.java
@@ -22,7 +22,7 @@ import io.kestra.core.tenant.TenantService;
 import io.kestra.core.utils.IdUtils;
 import io.kestra.core.utils.QueryFilterTestUtils;
 import io.kestra.core.utils.TestsUtils;
-import io.kestra.webserver.responses.PagedResults;
+import io.kestra.webserver.responses.CursorPagedResults;
 import io.kestra.webserver.tenants.TenantValidationFilter;
 
 import io.micronaut.core.type.Argument;
@@ -82,15 +82,15 @@ class LogControllerTest {
         logRepository.save(log2);
         logRepository.save(log3);
 
-        PagedResults<LogEntry> logs = client.toBlocking().retrieve(
+        CursorPagedResults<LogEntry> logs = client.toBlocking().retrieve(
             GET("/api/v1/" + tenant + "/logs/search"),
-            Argument.of(PagedResults.class, LogEntry.class)
+            Argument.of(CursorPagedResults.class, LogEntry.class)
         );
         assertThat(logs.getTotal()).isEqualTo(3L);
 
         logs = client.toBlocking().retrieve(
             GET("/api/v1/" + tenant + "/logs/search?filters[level][GREATER_THAN_OR_EQUAL_TO]=INFO"),
-            Argument.of(PagedResults.class, LogEntry.class)
+            Argument.of(CursorPagedResults.class, LogEntry.class)
         );
         assertThat(logs.getTotal()).isEqualTo(2L);
 
@@ -156,16 +156,16 @@ class LogControllerTest {
         assertThat(logs.getFirst().getExecutionKind()).isEqualTo(ExecutionKind.PLAYGROUND);
 
         // Global search defaults to NORMAL kind only too...
-        PagedResults<LogEntry> search = client.toBlocking().retrieve(
+        CursorPagedResults<LogEntry> search = client.toBlocking().retrieve(
             GET("/api/v1/" + tenant + "/logs/search"),
-            Argument.of(PagedResults.class, LogEntry.class)
+            Argument.of(CursorPagedResults.class, LogEntry.class)
         );
         assertThat(search.getTotal()).isEqualTo(0L);
 
         // ...and can be narrowed with an explicit KIND filter.
         search = client.toBlocking().retrieve(
             GET("/api/v1/" + tenant + "/logs/search?filters[kind][EQUALS]=PLAYGROUND"),
-            Argument.of(PagedResults.class, LogEntry.class)
+            Argument.of(CursorPagedResults.class, LogEntry.class)
         );
         assertThat(search.getTotal()).isEqualTo(1L);
     }
@@ -273,9 +273,9 @@ class LogControllerTest {
         logRepository.save(log2);
         logRepository.save(log3);
 
-        PagedResults<LogEntry> logs = client.toBlocking().retrieve(
+        CursorPagedResults<LogEntry> logs = client.toBlocking().retrieve(
             GET("/api/v1/" + tenant + "/logs/search?filters[timeRange][EQUALS]=PT25H"),
-            Argument.of(PagedResults.class, LogEntry.class)
+            Argument.of(CursorPagedResults.class, LogEntry.class)
         );
         assertThat(logs.getTotal()).isEqualTo(2L);
     }
